Apply review and little refactor
